SDA-5: request to api of edit profile fixed.

+++ app/src/main/java/sdumchykov/androidApp/presentation/signUp/CredentialsViewModel.kt
@@ -50,7 +50,7 @@ class CredentialsViewModel @Inject constructor(
                         AppDatabase::class.java, "database-name"
                     ).build()
                     val userDao = db.userDao()
-                    userDao.delete(userDao.getUser())
+                    if (userDao.getUser() != null) userDao.delete(userDao.getUser())
 
                     userDao.insert(
                         sdumchykov.androidApp.domain.local.User(
